+++ R/prediction_hetglm.R
@@ -1,0 +1,21 @@
+#' @rdname prediction
+#' @export
+prediction.hetglm <- function(model, data = find_data(model, parent.frame()), type = c("response", "link", "scale"), ...) {
+    # setup data
+    data <- data
+    
+    type <- match.arg(type)
+    
+    # extract predicted value at input value (value can only be 1 number)
+    pred <- data.frame(fitted = predict(model, newdata = data, type = type, ...),
+                       se.fitted = NA_real_)
+    class(pred[["fitted"]]) <- c("fit", "numeric")
+    class(pred[["se.fitted"]]) <- c("se.fit", "numeric")
+    
+    # obs-x-(ncol(data)+2) data.frame of predictions
+    structure(cbind(data, pred), 
+              class = c("prediction", "data.frame"), 
+              row.names = seq_len(length(pred[["fit"]])),
+              model.class = class(model),
+              type = type)
+}